+++ src/modules/PlayerBots/playerbot/strategy/triggers/TriggerContext.h
@@ -408,6 +408,8 @@ namespace ai
             creators["noth plagued adds nearby"] = [](PlayerbotAI* ai) { return new NothPlaguedAddsNearbyTrigger(ai); };
             creators["maexxna necrotic poison"] = [](PlayerbotAI* ai) { return new MaexxnaNecroticPoisonTrigger(ai); };
             creators["noth curse plaguebringer"] = [](PlayerbotAI* ai) { return new NothCursePlaguebringerTrigger(ai); };
+            creators["faerlina poison bolt"] = [](PlayerbotAI* ai) { return new FaerlinaPoisonBoltTrigger(ai); };
+            creators["ossirian curse of tongues"] = [](PlayerbotAI* ai) { return new OssirianCurseOfTonguesTrigger(ai); };
             creators["marli spawn nearby"] = [](PlayerbotAI* ai) { return new MarliSpawnNearbyTrigger(ai); };
             creators["jindo totems nearby"] = [](PlayerbotAI* ai) { return new JindoTotemsNearbyTrigger(ai); };
             creators["thekal zealots nearby"] = [](PlayerbotAI* ai) { return new ThekalZealotsNearbyTrigger(ai); };